adding the who is missing feature and changing the message when we complete hours
