+++ src/subscription.js
@@ -1,12 +1,11 @@
-const builder = require('botbuilder');
 const moment = require('moment-timezone');
 const business = require('moment-business');
-const fetch = require('node-fetch');
-const { find, pull, get } = require('lodash');
-const { bot } = require('./connector');
+const { find, pull } = require('lodash');
 const CONFIG = require('./config/env');
-const { log, trace, error } = require('./config/debug')(__filename);
+const { log, error } = require('./config/debug')(__filename);
 const Task = require('./task');
+const sendMessage = require('./util/sendMessage');
+const getHoliday = require('./util/getHoliday');
 
 /**
  * Store timeout references in memory
@@ -54,49 +53,7 @@ function scheduleMessage(now, nextDay, holiday, subscription) {
     removeSubscription(subscription);
 
     const task = new Task(() => {
-        const today = moment().tz('US/Central');
-        const lastBusinessDay = business.subtractWeekDays(today.clone(), 1);
-
-        if (holiday) {
-            while (holiday.isSame(lastBusinessDay, 'day')) {
-                log('Last business day was holiday. Going back one more day');
-                business.subtractWeekDays(lastBusinessDay, 1);
-            }
-        }
-
-        log('Last business day:', lastBusinessDay.format(CONFIG.DATE_FORMAT));
-
-        fetch(`${CONFIG.PENGUIN_REPORT_API}?date=${lastBusinessDay.format('YYYY-MM-DD')}`)
-        .then((res) => res.json())
-        .then((data) => {
-            trace('%O', data);
-
-            let message;
-
-            if (!data) {
-                message = 'Hubo un error al llamar al API de Penguin Report. 😅';
-            }
-
-            const penguined = data.filter((peep) => peep.totalHours < 7)
-            .map((peep) => `🐧  <b>${peep['person-name']}</b> (${peep.totalHours} horas)`);
-
-            if (penguined.length) {
-                message = 'Las siguientes personas aún no han reportado horas:\n\n' +
-                penguined.join('\n\n') +
-                '\n\n---\n\n' + CONFIG.PENGUIN_REPORT_URL +
-                '?date=' + lastBusinessDay.format('YYYY-MM-DD');
-            } else {
-                message = 'Todos reportaron horas. ¡Yay! 🙌';
-            }
-
-            const msg = new builder.Message()
-            .address(subscription)
-            .text('¡Buenos Días!\n\n' + message );
-
-            trace('%O', msg);
-
-            return bot.send(msg);
-        })
+        sendMessage(holiday, subscription)
         .then(() => {
             log('Programmed message sent!');
             addSubscription(subscription);
@@ -138,33 +95,17 @@ function addSubscription(subscription) {
     .second(Math.floor(Math.random() * 60))
     .millisecond(0);
 
-    fetch(`${CONFIG.HOLIDAYS_API_URL}/CO/${nextDay.year()}}`)
-    .then((res) => res.json())
-    .catch((err) => {
-        error(err);
-        return Promise.resolve([]);
-    })
-    .then((response) => {
-        trace(response);
-
-        const holidays = get(response, 'holidays');
-        let holiday;
-
-        if (holidays) {
-            holiday = holidays
-            .map((h) => moment(h.date))
-            .filter((h) => h.isSameOrAfter(now, 'day'))[0];
-
-            log('Next Colombian Holiday:', holiday.format(CONFIG.DATE_FORMAT));
-
-            while (holiday.isSame(nextDay, 'day')) {
+    getHoliday(nextDay)
+    .then((holiday) => {
+       if (holiday) {
+           while (holiday.isSame(nextDay, 'day')) {
                 log('Adding one more day to next day count');
 
                 business.addWeekDays(nextDay, 1);
             }
-        }
+       }
 
-        scheduleMessage(now, nextDay, holiday, subscription);
+       scheduleMessage(now, nextDay, holiday, subscription);
     });
 }
